xkiss16 conv fix 3

+++ generators/xkiss16sh_awc.c
@@ -80,7 +80,7 @@ static inline uint16_t Xkiss16AwcState_get_bits(Xkiss16AwcState *obj)
     uint16_t awc = rotl16(obj->awc_x0, 3) ^ (obj->awc_x1);
     const uint16_t out = (uint16_t) (obj->xs + obj->weyl);
     const int rot = obj->awc_x0 & 0xF;
-    return awc + rotl16(out, rot);
+    return (uint16_t) (awc + rotl16(out, rot));
 //    return rotl16(awc, obj->xs) + rotl16(obj->xs + obj->weyl, obj->awc_x1 & 0xF);
 }
